requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
supabase==2.3.1
numpy==1.26.4
altair==5.2.0
streamlit-extras==0.3.0 